requirements.txt
tensorflow
numpy
pillow
scikit-learn
matplotlib
tqdm
nltk
pycocoevalcap
opencv-python